make interfaces Clock\* final classes
