--- src/Earth/Clock/Month.php
@@ -1,48 +1,0 @@
-<?php
-declare(strict_types = 1);
-
-namespace Innmind\TimeContinuum\Earth\Clock;
-
-use Innmind\TimeContinuum\{
-    Clock\Month as MonthInterface,
-    Exception\DomainException,
-};
-
-/**
- * @psalm-immutable
- */
-final class Month implements MonthInterface
-{
-    private int $month;
-    private int $days;
-    private string $string;
-
-    public function __construct(Year $year, int $month)
-    {
-        if ($month < 1 || $month > 12) {
-            throw new DomainException((string) $month);
-        }
-
-        $this->month = $month;
-        $this->days = (int) \date(
-            't',
-            $time = \mktime(0, 0, 0, $month, 1, $year->toInt()),
-        );
-        $this->string = \date('F', $time);
-    }
-
-    public function numberOfDays(): int
-    {
-        return $this->days;
-    }
-
-    public function toInt(): int
-    {
-        return $this->month;
-    }
-
-    public function toString(): string
-    {
-        return $this->string;
-    }
-}
