+++ tests/Clock/YearTest.php
@@ -1,26 +1,22 @@
 <?php
 declare(strict_types = 1);
 
-namespace Tests\Innmind\TimeContinuum\Earth\Clock;
+namespace Tests\Innmind\TimeContinuum\Clock;
 
-use Innmind\TimeContinuum\{
-    Earth\Clock\Year,
-    Clock\Year as YearInterface,
-};
+use Innmind\TimeContinuum\Clock\Year;
 use PHPUnit\Framework\TestCase;
 
 class YearTest extends TestCase
 {
     public function testInterface()
     {
-        $year = new Year(2016);
+        $year = Year::of(2016);
 
-        $this->assertInstanceOf(YearInterface::class, $year);
         $this->assertSame(366, $year->numberOfDays());
         $this->assertSame(12, $year->numberOfMonths());
         $this->assertSame(2016, $year->toInt());
         $this->assertSame('2016', $year->toString());
 
-        $this->assertSame(365, (new Year(2017))->numberOfDays());
+        $this->assertSame(365, Year::of(2017)->numberOfDays());
     }
 }
